examples/training/dbrx/requirements.txt
transformers==4.48.0
pytorch-lightning==2.5.0
regex
tensorboard
datasets
sentencepiece
